### 2020-03-06 #2148 @whedon
[PRE REVIEW]: Spleeter: a fast and state-of-the art music source separation tool with pre-trained models
**Submitting author:** @romi1502 (<a href="http://orcid.org/0000-0001-8158-5562">Romain Hennequin</a>)
**Repository:** <a href="https://github.com/deezer/spleeter" target ="_blank">https://github.com/deezer/spleeter</a>
**Version:** v1.4.9
