### 2020-03-07 #2148 @terrytangyuan
Happy to edit this one. @romi1502 if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.
