### 2020-03-09 #2148 @romi1502
Hey @terrytangyuan.
Thank you for managing the review.
An obvious choice for the paper review would be faroit as he is the main author of an open source package which also deals with Source Separation (Open Unmix) though it may be considered as a conflict of interest. Otherwise, I don't know other people in the list. Anyone with experience in Python and audio/music processing should be skilled to review it.
