### 2020-03-20 2365a9b
Removing duplicate logs
