### 2020-03-20 6539bf7
Bumping tensorflow version to do 1 PR merge instead of two
