### 2020-03-26 5d18b78
Adding support for mono files with librosa backend
