### 2020-03-27 38bfff8
Fixing tests when running in single process
