### 2020-03-27 f8c728a
Hack so that multichannel doesn't crash with librosa
