### 2020-03-27 deezer/spleeter#307 @bmcfee
[Discussion] JOSS review point: installation instructions
Hey spleeter devs, I'm a reviewer on your (JOSS submission](https://github.com/openjournals/joss-reviews/issues/2154)

Working down the checklist, I had a question about your [installation instructions](https://github.com/deezer/spleeter#quick-start).  They appear to be correct (ie it works), but I'm confused by the instruction to both clone the repo and install from conda-forge.  I assume this is only written this way to ensure that the demo track is available for running the test command.

I might suggest breaking this up into more explicit sections for installing the package (for use), installing for development (git clone), and testing on an example audio file.
