### 2020-03-29 deezer/spleeter#307 @mmoussallam
Hi @bmcfee

First of all, thanks a lot for taking upon your time to review our submission, especially considering the worldwide situation going on.

Indeed you're right that the cloning is not really needed now. In a first version we were also locally building a `conda` env from a cloned file. But since we've ported spleeter to conda-forge we should probably update the documentation accordingly.

I'm letting this issue open until we do so.

Stay safe and I hope you'll have some fun in using spleeter.
