### 2020-04-01 deezer/spleeter#307 @romi1502
Hi @bmcfee
Thank you for the suggestion. Indeed, it was unclear that the `git clone` was only for downloading the audio demo file.
We updated the readme for making the `Quick start instructions` more explicit by adding comments and replacing the `git clone` by a `wget`.
Let us know if you feel it may be improved further.
