### 2020-04-10 a23a66e
Cleaning: updated docstring and unexposed some internal methods
