### 2020-04-14 87225b9
Put back save_to_file as a public method of Separator
