### 2020-04-26 0366a4c
Use `with statement` to auto-close tarfile
