### 2020-04-26 ba9999a
Remove temp model archive after extraction
Without this change every time a model is downloaded
it leaks a file in the temp directory.
