+++ spleeter/model/__init__.py
@@ -218,22 +218,22 @@ class EstimatorSpecBuilder(object):
         loss_type = self._params.get('loss_type', self.L1_MASK)
         if loss_type == self.L1_MASK:
             losses = {
-                name: tf.reduce_mean(tf.abs(output - labels[name]))
+                name: tf.reduce_mean(input_tensor=tf.abs(output - labels[name]))
                 for name, output in output_dict.items()
             }
         elif loss_type == self.WEIGHTED_L1_MASK:
             losses = {
                 name: tf.reduce_mean(
-                    tf.reduce_mean(
-                        labels[name],
+                    input_tensor=tf.reduce_mean(
+                        input_tensor=labels[name],
                         axis=[1, 2, 3],
-                        keep_dims=True) *
+                        keepdims=True) *
                     tf.abs(output - labels[name]))
                 for name, output in output_dict.items()
             }
         else:
             raise ValueError(f"Unkwnown loss type: {loss_type}")
-        loss = tf.reduce_sum(list(losses.values()))
+        loss = tf.reduce_sum(input_tensor=list(losses.values()))
         # Add metrics for monitoring each instrument.
         metrics = {k: tf.compat.v1.metrics.mean(v) for k, v in losses.items()}
         metrics['absolute_difference'] = tf.compat.v1.metrics.mean(loss)
@@ -276,8 +276,8 @@ class EstimatorSpecBuilder(object):
 
         if stft_name not in self._features:
             stft_feature = tf.transpose(
-                stft(
-                    tf.transpose(self._features['waveform']),
+                a=stft(
+                    tf.transpose(a=self._features['waveform']),
                     self._frame_length,
                     self._frame_step,
                     window_fn=lambda frame_length, dtype: (
@@ -332,15 +332,15 @@ class EstimatorSpecBuilder(object):
         :returns: inverse STFT (waveform)
         """
         inversed = inverse_stft(
-            tf.transpose(stft_t, perm=[2, 0, 1]),
+            tf.transpose(a=stft_t, perm=[2, 0, 1]),
             self._frame_length,
             self._frame_step,
             window_fn=lambda frame_length, dtype: (
                 hann_window(frame_length, periodic=True, dtype=dtype))
         ) * self.WINDOW_COMPENSATION_FACTOR
-        reshaped = tf.transpose(inversed)
+        reshaped = tf.transpose(a=inversed)
         if time_crop is None:
-            time_crop = tf.shape(self._features['waveform'])[0]
+            time_crop = tf.shape(input=self._features['waveform'])[0]
         return reshaped[:time_crop, :]
 
     def _build_mwf_output_waveform(self):
@@ -359,7 +359,7 @@ class EstimatorSpecBuilder(object):
                 pad_and_reshape(
                     output_dict[f'{instrument}_spectrogram'],
                     self._frame_length,
-                    self._F)[:tf.shape(x)[0], ...]
+                    self._F)[:tf.shape(input=x)[0], ...]
                 for instrument in self._instruments
             ],
             axis=3)
@@ -385,11 +385,11 @@ class EstimatorSpecBuilder(object):
         # Extend with average
         # (dispatch according to energy in the processed band)
         if extension == "average":
-            extension_row = tf.reduce_mean(mask, axis=2, keepdims=True)
+            extension_row = tf.reduce_mean(input_tensor=mask, axis=2, keepdims=True)
         # Extend with 0
         # (avoid extension artifacts but not conservative separation)
         elif extension == "zeros":
-            mask_shape = tf.shape(mask)
+            mask_shape = tf.shape(input=mask)
             extension_row = tf.zeros((
                 mask_shape[0],
                 mask_shape[1],
[…]
