### 2020-05-03 7afbeb1
add stempeg adapter
