+++ spleeter/audio/ffmpeg.py
@@ -11,7 +11,7 @@
 import os
 
 # pylint: disable=import-error
-import ffmpeg
+import stempeg
 import numpy as np
 # pylint: enable=import-error
 
@@ -24,26 +24,7 @@ __author__ = 'Deezer Research'
 __license__ = 'MIT License'
 
 
-def _to_ffmpeg_time(n):
-    """ Format number of seconds to time expected by FFMPEG.
-    :param n: Time in seconds to format.
-    :returns: Formatted time in FFMPEG format.
-    """
-    m, s = divmod(n, 60)
-    h, m = divmod(m, 60)
-    return '%d:%02d:%09.6f' % (h, m, s)
-
-
-def _to_ffmpeg_codec(codec):
-    ffmpeg_codecs = {
-        'm4a': 'aac',
-        'ogg': 'libvorbis',
-        'wma': 'wmav2',
-    }
-    return ffmpeg_codecs.get(codec) or codec
-
-
-class FFMPEGProcessAudioAdapter(AudioAdapter):
+class StempegProcessAudioAdapter(AudioAdapter):
     """ An AudioAdapter implementation that use FFMPEG binary through
     subprocess in order to perform I/O operation for audio processing.
 
@@ -66,37 +47,16 @@ class FFMPEGProcessAudioAdapter(AudioAdapter):
         :returns: Loaded data a (waveform, sample_rate) tuple.
         :raise SpleeterError: If any error occurs while loading audio.
         """
-        if not isinstance(path, str):
-            path = path.decode()
-        try:
-            probe = ffmpeg.probe(path)
-        except ffmpeg._run.Error as e:
-            raise SpleeterError(
-                'An error occurs with ffprobe (see ffprobe output below)\n\n{}'
-                .format(e.stderr.decode()))
-        if 'streams' not in probe or len(probe['streams']) == 0:
-            raise SpleeterError('No stream was found with ffprobe')
-        metadata = next(
-            stream
-            for stream in probe['streams']
-            if stream['codec_type'] == 'audio')
-        n_channels = metadata['channels']
-        if sample_rate is None:
-            sample_rate = metadata['sample_rate']
-        output_kwargs = {'format': 'f32le', 'ar': sample_rate}
-        if duration is not None:
-            output_kwargs['t'] = _to_ffmpeg_time(duration)
-        if offset is not None:
-            output_kwargs['ss'] = _to_ffmpeg_time(offset)
-        process = (
-            ffmpeg
-            .input(path)
-            .output('pipe:', **output_kwargs)
-            .run_async(pipe_stdout=True, pipe_stderr=True))
-        buffer, _ = process.communicate()
-        waveform = np.frombuffer(buffer, dtype='<f4').reshape(-1, n_channels)
-        if not waveform.dtype == np.dtype(dtype):
-            waveform = waveform.astype(dtype)
+        waveform, sample_rate = stempeg.read_streams(
+            path=path,
+            start=offset,
+            duration=duration,
+            stem_id=None,
+            dtype=dtype,
+            info=None,
+            sample_rate=sample_rate,
+            stems_from_multichannel=False
+        )
         return (waveform, sample_rate)
 
     def save(
@@ -112,26 +72,12 @@ class FFMPEGProcessAudioAdapter(AudioAdapter):
         :param bitrate: (Optional) Bitrate of the written audio file.
         :raise IOError: If any error occurs while using FFMPEG to write data.
         """
-        directory = os.path.dirname(path)
-        if not os.path.exists(directory):
-            raise SpleeterError(f'output directory does not exists: {directory}')
         get_logger().debug('Writing file %s', path)
-        input_kwargs = {'ar': sample_rate, 'ac': data.shape[1]}
[…]
