### 2020-05-07 9513380
Use tf.abs in compute_spectrogram_tf
To avoid the following error: `Cannot convert a symbolic Tensor (transpose_1:0) to a numpy array`
