### 2020-05-07 e2a68f2
docs: add JOSS badge
