### 2020-05-10 1c1ff80
Update README.md
Added a disclaimer line to avoid questions on forked initiatives
