### 2020-05-12 0f89b71
Avoid multiple checkpoint restauration and add a specific tf graph for each Separator which makes it possible to instanciate several ones
