### 2020-05-19 deezer/spleeter#381 @faroit
[JOSS] add demucs SIR, SAR, ISR scores
## Description

![image](https://user-images.githubusercontent.com/72940/82384986-55190700-9a31-11ea-9320-7da00baa155e.png)

the performance comparison section in the joss paper does not show the scores for demucs on SIR, ISR and SAR. These scores seems to be easily available using the scripts from the demucs repo (see [here](https://github.com/facebookresearch/demucs#examining-the-results-from-the-paper-experiments).

To speed this up, maybe @adefossez could quickly provide these numbers here.

-- this issue is part of [a JOSS review](https://github.com/openjournals/joss-reviews/issues/2154)
