### 2020-05-19 deezer/spleeter#381 @faroit
@adefossez Perfect, thanks.

The authors of Spleeter have submitted a paper to the Journal of Open-Source Software. It's a really great experience.
