### 2020-05-20 08b70d5
Fix typo in README.md
Seems appropriate enough that my first contribution would be the "contributing" section of the readme...
