### 2020-05-20 deezer/spleeter#381 @romi1502
@adefossez, thank you for the results. Seems I was too lazy to read until the appendix when I reported the numbers :). By the way, I can't find the ISR in the paper. Do you have them somewhere?
@faroit, thank you for pointing out this issue. I just pushed the SIR and SAR of Demucs in [this commit](https://github.com/deezer/spleeter/compare/4b8b9236d08b...096d9ce19332).
