### 2020-05-20 d8c1e68
[joss] spelling and grammar fixes
