### 2020-05-20 #2154 @faroit
# Review/Comments

Spleeter is a very valuable addition to the music separation ecosystem. The software is already hugely popular. The majority of its users are end-users without scientific background, so that it can be said that spleeter has made our research domain significantly more popular, which is a great achievement for a software package. A paper about it here, hence naturally deserves publication.

I see two reasons for the success of spleeter:
    1. The code is thoughtfully designed and sufficiently fast to run inference on simple desktop clients.
    2. Separation performance _with the provided pre-trained models_ is very good, which allows users to separate audio simply and nicely.

Nonetheless, I want to bring up two issues that I see with respect to the performance as stated in the paper.

## 1. Reproducibility

Although it may definitely be helpful as a pre-processing step in some domains, I think it is fair to say that Spleeter does not contribute significantly to the advance of source separation research _per se_. This is mainly so because its good performance comes from the fact that it was trained on a private dataset, that was not made public, and that it actually behaves not as good when trained on the widely used MUSDB18, at least [with the provided configuration files](https://github.com/deezer/spleeter/issues/81).

This fact prevents other researchers from reproducing results. While the authors are very clear about this fact in their paper, it would be very easy to at least report the results obtained on __MUSDB18__, and tune a training configuration that would be optimized for this case, as many other source separation researchers are doing. This would allow other researchers to decide on comparable grounds whether they want to use spleeter for research or another system.

## 2. "State-of-the-art" performance

Given that other recent architectures such as the ones listed [here](https://paperswithcode.com/sota/music-source-separation-on-musdb18) surpassed the performance of spleeter, I think it is be slightly miss-leading to keep `state-of-the-art` in the title of this paper.

All in all, I would see these two points as a starting point for discussion with the other reviewers (@bmcfee) and @terrytangyuan, and the authors. In the meantime, maybe issue [#381](https://github.com/deezer/spleeter/issues/381) and [#384](https://github.com/deezer/spleeter/issues/384) can be addressed.
