### 2020-05-22 80250e7
Added an extra 'foldername' option in name formatting and use this option to fix eval
