### 2020-05-22 deezer/spleeter#381 @romi1502
Ok, I also added the ISR in the paper in [this commit](https://github.com/deezer/spleeter/compare/096d9ce19332...bd9cd3e00658). So all four metrics are now in the paper for demucs.
I close the issue then.
