### 2020-05-23 deezer/spleeter#393 @bmcfee
[Discussion] [JOSS Review] Community guidelines
Hey everyone, picking up the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2154) again. (Sorry for the lag, but you know, 2020...)  |

Working on this point:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I noticed that you do have a CONTRIBUTING.MD file, but it's not linked from the main README, and it takes a bit of digging on github / wiki to get there.  Since it's hidden under the `.githhub/` folder, I don't think you can expect new users to find it immediately.  I suggest linking this directly from the readme file  in the contributing section to make it easier to find.

Bugs and support are fine as is.

Two more related points:

1. README mentions that it's MIT licensed, but does not link to the license file.  (Minor thing, but I see no reason not to.)
2. CONTRIBUTING.MD is pretty minimal -- it wouldn't hurt to replicate the development / testing instructions here, and add a bit more about creating branches and sending a pull request.
