### 2020-05-23 #2154 @bmcfee
Thanks @faroit for laying this out carefully.  I agree with point 2.

Point 1 is a bit trickier, and I can see both sides of the issue here. (This is why I've left the "performance" box unchecked for now.)

- On the one hand, the authors are making an existential claim that their pre-trained model achieves the reported scores on musdb.  I haven't verified this exactly, but in working with this package over the past few months, I don't see it as controversial.
- On the other hand, one could argue that the contribution here is the entire implementation, including training algorithm, and the pre-trained model parameters are just one (important) example produced by their system on proprietary data.  As @faroit correctly states, this does not guarantee that retraining on open data will produce the same results (and nobody expects that), but there is no reported benchmark for that condition to compare against anyway.

So to me, the question is: are we evaluating the pre-trained model (the application), or the entire framework which produced it?  Put more succinctly: is training within scope for the review or not?  If so, then we should have some benchmarks on open data to verify the results (even if they're below what's reported with the private training set).  If not, I'm fine to approve it as is.  But I think we need some editorial guidance here -- @terrytangyuan ?
