### 2020-05-23 #2154 @faroit
@bmcfee Just adding to your comments that point 1 was addressed yesterday as documented [here](https://github.com/deezer/spleeter/issues/81#issuecomment-632769524).

That means there are settings available now for reproducible training on publicly available data. Even though the scores are significantly below SOTA, I still see them as very valuable for other researchers. And I therefore think that Point 1 would be fully addressed if these scores are stated in either of the following form

* in the paper, or
* in the spleeter docs/wiki, or
* directly commented in the [json config file](https://github.com/deezer/spleeter/blob/master/configs/musdb_config.json)
