### 2020-05-23 #2154 @bmcfee
Oh great, I hadn't seen that yet.  I would vote for paper + docs/wiki.
