### 2020-05-24 deezer/spleeter#393 @mmoussallam
Hi @bmcfee

Thanks for the suggestions, I've updated the README and CONTRIBUTING files to take that into account.
