### 2020-05-25 a27aa96
Linking to the separation result page
