### 2020-05-26 #2154 @romi1502
Thanks @faroit and @bmcfee for your comments.

Issues [#381](https://github.com/deezer/spleeter/issues/381) and [#384](https://github.com/deezer/spleeter/issues/381) have been addressed and closed.

Regarding the 2 aspects mentioned by @faroit:

> 1. Reproducibility

We did the following modification:
- We've modified the [config file](https://github.com/deezer/spleeter/blob/master/configs/musdb_config.json) that makes it possible to train spleeter on the musdb dataset and it trains properly now.
- We've added [this page](https://github.com/deezer/spleeter/wiki/Separation-Performances) in the wiki to talk about performances of spleeter with a paragraph on a model trained on the musdb train set that reports a performance table (we also put the command to reproduce the training of the model). We also added a link to it in the [README](https://github.com/deezer/spleeter/blob/master/README.md) of the repo.

So it should be quite easy now to reproduce these results for anyone.

However, we think that putting an extra table for a model trained on a different dataset in the paper may bring confusion. As already stated, the added value of spleeter mainly comes from the provided pretrained models that have quite good performances (because they were trained on a private dataset): as the paper is not targeted at the source separation community but rather at MIR researchers needing a simple and efficient tool to perform separation as a pre-processing, putting a highlight on models trained on musdb in the JOSS paper (which is supposed to be concise) may sound confusing.

> 2. "State-of-the-art" performance

We agree that the use of SOA may be misleading (but advertising ;) ).
We propose to change the title of the paper to "Spleeter: a fast and efficient music source separation tool with pre-trained models" to keep a quite similar one without the SOA claim.

What do you think about these modifications/suggestions?
