### 2020-05-28 4349047
Replaced SOA by efficient in paper title
