### 2020-05-28 #2154 @romi1502
Ok,
Thanks @faroit and @bmcfee for your reviews.
@terrytangyuan, I think we're good to move forward.
We will just need to change the title of the paper: I already updated the [paper.md](https://github.com/deezer/spleeter/blob/master/paper.md) of the project repo, but we'll need to update the submission title too. Do you how we can do it ?
