### 2020-06-02 8af5ee2
missing command to switch to the created directory
