### 2020-06-05 e22c64a
Add test for evaluation
