### 2020-06-05 ff3ac62
Fix fake data generation
