### 2020-06-05 582cb34
fix: add musdb test dev for CI
