+++ .circleci/config.yml
@@ -12,7 +12,7 @@ jobs:
       - restore_cache:
           key: models-{{ checksum "spleeter/model/__init__.py" }}
       - run: apt-get update && apt-get install -y ffmpeg
-      - run: pip install -r requirements.txt && pip install pytest pytest-xdist
+      - run: pip install -r requirements.txt && pip install pytest pytest-xdist musdb
       - run: make test
       - save_cache:
           key: models-{{ checksum "spleeter/model/__init__.py" }}
@@ -30,7 +30,7 @@ jobs:
       - restore_cache:
           key: models-{{ checksum "spleeter/model/__init__.py" }}
       - run: apt-get update && apt-get install -y ffmpeg
-      - run: pip install -r requirements.txt && pip install pytest pytest-xdist
+      - run: pip install -r requirements.txt && pip install pytest pytest-xdist musdb
       - run: make test
       - save_cache:
           key: models-{{ checksum "spleeter/model/__init__.py" }}
