### 2020-06-05 9dcf220
fix: add museval dep for CI testing
