### 2020-06-17 3e4a644
fixed numba version for fixing librosa dependency issue
