### 2020-06-17 d9e4834
fixed numba version for fixing librosa dependency issue
