### 2020-06-18 8d442c8
Fixing gltches issues with Istft
