### 2020-06-18 bec0555
remove useless prints
