### 2020-06-18 4b40685
remove useless prints
