### 2020-06-18 a28665c
fix: conda-gpu base image
