### 2020-06-18 02f86e4
investigate diff values
