### 2020-06-18 dfbc579
fix: spleeter GPU package
