### 2020-06-18 05444e4
fix: check miniconda base
