### 2020-06-18 e6f6422
fix: restore conda model
