### 2020-06-18 af6dbec
fix tests and more testing
