### 2020-06-19 b8e4d21
refactor: switch gpu tag to 3.7-gpu
