### 2020-06-24 #2154 @romi1502
Thank you @arfon
The latest spleeter version (1.5.3) we've just released includes all changes discussed in the review.
I've just uploaded an archive with this version on [Zenodo](https://zenodo.org/record/3906389) with same title and same authors as the paper.

It has DOI **10.5281/zenodo.3906389**.

Let us know if we need to anything else.
