### 2020-06-24 #2154 @arfon
> - https://doi.org/10.1007/978-3-319-93764-9_28 may be missing for title: The 2018 Signal Separation Evaluation Campaign

@romi1502 - can you check if this ☝️  DOI is correct? If it is, please add it to your BibTeX file.
