### 2020-06-24 7f320f6
Updated bibliography with DOI
