+++ paper.bib
@@ -1,19 +1,13 @@
 % bibtex
 
-@article{SISEC18,
+@inproceedings{SISEC18,
        author = {{St{\"o}ter}, Fabian-Robert and {Liutkus}, Antoine and {Ito}, Nobutaka},
-        title = "{The 2018 Signal Separation Evaluation Campaign}",
-      journal = {arXiv e-prints},
-     keywords = {Electrical Engineering and Systems Science - Audio and Speech Processing, Computer Science - Sound},
-         year = "2018",
-        month = "Apr",
-          eid = {arXiv:1804.06267},
-        pages = {arXiv:1804.06267},
-archivePrefix = {arXiv},
-       eprint = {1804.06267},
- primaryClass = {eess.AS},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2018arXiv180406267S},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+        title = {The 2018 Signal Separation Evaluation Campaign},
+        year = {2018},
+        booktitle = {Latent Variable Analysis and Signal Separation. LVA/ICA},
+        vol={10891},
+        doi = {10.1007/978-3-319-93764-9_28},
+        publisher = { Springer, Cham}
 }
 
 @misc{spleeter2019,
