### 2020-06-24 48f5e01
Force capitalization in bibtex
