+++ paper.bib
@@ -4,7 +4,7 @@
        author = {{St{\"o}ter}, Fabian-Robert and {Liutkus}, Antoine and {Ito}, Nobutaka},
         title = {The 2018 Signal Separation Evaluation Campaign},
         year = {2018},
-        booktitle = {Latent Variable Analysis and Signal Separation. LVA/ICA},
+        booktitle = {Latent Variable Analysis and Signal Separation. {LVA}/{ICA}},
         vol={10891},
         doi = {10.1007/978-3-319-93764-9_28},
         publisher = { Springer, Cham}
