### 2020-06-24 #2154 @romi1502
We were actually citing the arxiv version of the paper which is not good. I replaced it by the published version and added the DOI in the BibTex file.
