### 2020-06-24 #2154 @faroit
@romi1502 @arfon My first experience with JOSS when I submitted a paper was excellent and I really tried to help so that other submissions like this one could get the same treatment but unfortunately this was really difficult during COVID.
So thanks again for your patience!
